added parkingSlotStrategy

+++ practice/parkinglot/models/payments/CardPaymentStrategy.java
@@ -3,7 +3,7 @@ package Low_Level_Design.practice.parkinglot.models.payments;
 public class CardPaymentStrategy implements PaymentStrategy {
 
     @Override
-    public void pay(int amount) {
+    public void pay(float amount) {
         System.out.println("Paid via Card");
     }
