+++ practice/parkinglot/models/payments/FastTagPaymentStrategy.java
@@ -3,7 +3,7 @@ package Low_Level_Design.practice.parkinglot.models.payments;
 public class FastTagPaymentStrategy implements PaymentStrategy{
 
     @Override
-    public void pay(int pay) {
+    public void pay(float pay) {
         System.out.println("Paid by fast tag");
     }
