+++ practice/parkinglot/models/payments/PaymentStrategy.java
@@ -1,6 +1,6 @@
 package Low_Level_Design.practice.parkinglot.models.payments;
 
 public interface PaymentStrategy {
-    void pay(int amount);
+    void pay(float amount);
     
 }
